+++ src/renderer/components/PortForwarding/Drawer/Template/index.js
@@ -1,0 +1,13 @@
+import titlePage from './TitlePage'
+import directionPage from './Body/DirectionPage'
+import sourcePage from './Body/SourcePage'
+import serverPage from './Body/ServerPage'
+import destinationPage from './Body/DestinationPage'
+
+export {
+    titlePage,
+    directionPage,
+    sourcePage,
+    serverPage,
+    destinationPage
+}
