--- src/renderer/components/PortForwarding/Drawer/components/index.js
@@ -1,11 +1,0 @@
-import directionPage from './DirectionPage'
-import sourcePage from './SourcePage'
-import serverPage from './ServerPage'
-import destinationPage from './DestinationPage'
-
-export {
-    directionPage,
-    sourcePage,
-    serverPage,
-    destinationPage
-}
