+++ src/renderer/components/PortForwarding/index.js
@@ -1,7 +1,7 @@
 import Title from './Title'
 import WelcomeBody from './Content/WelcomeBody'
 import SessionBody from './Content/TunnelingBody'
-import SessionDrawer from './Drawer/SessionDrawer'
+import SessionDrawer from './Drawer/NewSessionDrawer'
 
 export {
     Title,
